// frontend/packages/shared/src/api/models/campus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 校园轻享集市系统 API
 * ## 校园轻享集市系统 RESTful API 文档  基于 Java 21 + Spring Boot 3.x 的企业级单体应用  ### 主要功能 - 用户认证授权（JWT + RBAC） - 二手物品交易 - 订单管理与支付 - 即时通讯（WebSocket） - 论坛社区 - 积分与优惠券系统  ### 技术栈 - Java 21 (Virtual Threads) - Spring Boot 3.2.0 - PostgreSQL 16 - Redis 7.x - JWT 认证  ### 认证说明 大部分接口需要在请求头中携带 JWT Token： ``` Authorization: Bearer <your-jwt-token> ```  ### 获取 Token 1. 调用 `/auth/register` 注册账号 2. 调用 `/auth/login` 登录获取 Token 3. 在后续请求中使用该 Token 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: basui@campus.edu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 业务数据载荷
 * @export
 * @interface Campus
 */
export interface Campus {
    /**
     * 
     * @type {number}
     * @memberof Campus
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Campus
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof Campus
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Campus
     */
    'deleted'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Campus
     */
    'deletedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof Campus
     */
    'code'?: string;
    /**
     * 
     * @type {string}
     * @memberof Campus
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Campus
     */
    'status'?: CampusStatusEnum;
}

export const CampusStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;

export type CampusStatusEnum = typeof CampusStatusEnum[keyof typeof CampusStatusEnum];